dino-customdensefpnunetdecoder/requirements.txt
# Requirements for mecup_unsupervised

# Core - use available versions
torch>=2.6.0
torchvision>=0.16.0

# Vision transformers
timm>=0.9.12

# Image processing
opencv-python>=4.8.0
albumentations>=1.3.1
pillow>=10.0.0

# Numerical
numpy>=1.24.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Configuration
pyyaml>=6.0

# Progress
tqdm>=4.66.0

# ML utilities
scikit-learn>=1.3.0

# Logging
tensorboard>=2.15.0